+++ tests/http/test_oauth.py
@@ -1,7 +1,10 @@
 import base64
+import dataclasses
 import time
+from typing import Tuple
 
 import mock
+import pytest
 import requests
 import requests_mock as req_mock
 
@@ -11,8 +14,8 @@ HTTP_TIMEOUT = 10
 TOKEN_VALUE = "some-token"
 TOKEN_URL = "https://localhost/oauth/token"
 project_credentials = config.ProjectCredentialsSettings(
-    client_id="client_id",
-    client_secret="client_secret",
+    project_id="some-id",
+    project_api_key="some-key",
 )
 account_credentials = config.AccountCredentialsSettings(username="some-user", password="some-pass")
 
@@ -22,7 +25,7 @@ def basic_auth(username, password):
     return f'Basic {token.decode("ascii")}'
 
 
-basic_client_auth = basic_auth(project_credentials.client_id, project_credentials.client_secret)
+basic_client_auth = basic_auth(project_credentials.project_id, project_credentials.project_api_key)
 basic_auth_headers = {"Authorization": basic_client_auth}
 
 
@@ -31,7 +34,7 @@ class TestProjectTokenRetriever:
         def match_request_body(request):
             return request.text == "grant_type=client_credentials"
 
-        retrieve = oauth.ProjectTokenRetriever(lambda: project_credentials)
+        retrieve = oauth.ProjectTokenRetriever(project_credentials)
         requests_mock.post(
             TOKEN_URL,
             json={"access_token": TOKEN_VALUE},
@@ -51,7 +54,7 @@ class TestAccountTokenRetriever:
                 f"password={account_credentials.password}"
             )
 
-        retrieve = oauth.AccountTokenRetriever(lambda: account_credentials)
+        retrieve = oauth.AccountTokenRetriever(account_credentials)
         requests_mock.post(
             TOKEN_URL,
             json={"access_token": TOKEN_VALUE},
@@ -94,3 +97,40 @@ class TestUp42Auth:
         assert up42_auth.token.access_token == second_token
         assert TOKEN_URL, HTTP_TIMEOUT == retrieve.call_args.args[1:]
         assert retrieve.call_count == 2
+
+
+class TestDetectSettings:
+    def test_should_detect_project_credentials(self):
+        assert oauth.detect_settings(dataclasses.asdict(project_credentials)) == project_credentials
+
+    def test_should_detect_account_credentials(self):
+        assert oauth.detect_settings(dataclasses.asdict(account_credentials)) == account_credentials
+
+    @pytest.mark.parametrize("keys", [("project_id", "project_api_key"), ("username", "password")])
+    def test_should_accept_empty_credentials(self, keys: Tuple[str, str]):
+        credentials = dict(zip(keys, [None] * 2))
+        assert not oauth.detect_settings(credentials)
+
+    def test_fails_if_credentials_are_incomplete(self):
+        credentials = {"key1": "value1", "key2": None}
+        with pytest.raises(oauth.IncompleteCredentials):
+            oauth.detect_settings(credentials)
+
+    def test_fails_if_credentials_are_invalid(self):
+        credentials = {"key1": "value1", "key2": "value2"}
+        with pytest.raises(oauth.InvalidCredentials):
+            oauth.detect_settings(credentials)
+
+
+class TestDetectRetriever:
+    def test_should_detect_project_retriever(self):
+        assert isinstance(oauth.detect_retriever(project_credentials), oauth.ProjectTokenRetriever)
+
+    def test_should_detect_account_retriever(self):
+        assert isinstance(oauth.detect_retriever(account_credentials), oauth.AccountTokenRetriever)
+
+    def test_fails_if_settings_are_not_recognized(self):
+        credentials = mock.MagicMock()
+        with pytest.raises(oauth.UnsupportedSettings) as exc_info:
+            oauth.detect_retriever(credentials)
+        assert str(credentials) in str(exc_info.value)